requirements.txt
gradio>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0